+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog for NeoFS Node
 - SN now supports `ObjectService` requests, objects and session/bearer tokens signed using N3 account scheme (#3303, #3305)
 - SN attempts to read containers/eACLs calling new Container contract's `getContainerData`/`getEACLData` methods (#3323)
 - `request create-container` command to NeoFS CLI (#3319)
+- `max_valid_until_block_increment` setting to IR consensus configuration (#3335)
 
 ### Fixed
 - Bearer token signed not by its issuer is no longer passed (#3216)
@@ -65,7 +66,7 @@ Changelog for NeoFS Node
 ### Updated
 - `github.com/nspcc-dev/neofs-sdk-go` dependency to `v1.0.0-rc.13.0.20250417140404-8d69cb0e9a25` (#3255)
 - `github.com/nspcc-dev/neofs-contract` dependency to `v0.22.0` (#3282, #3305, #3323)
-- `github.com/nspcc-dev/neo-go` dependency to `v0.108.2-0.20250414115617-823a05fcc10b` (#3298)
+- `github.com/nspcc-dev/neo-go` dependency to `v0.109.0` (#3298, #3335)
 
 ### Updating from v0.45.2
 `apiclient.allow_external` field must be dropped from any SN configuration.